--- packages/logger/src/Providers/index.ts
@@ -1,3 +1,0 @@
-// Providers
-export * from './JsonProvider';
-export * from './ConsoleProvider';
