+++ packages/logger/tsdown.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'tsdown';
+import defaultConfig from '../../tsdown.config';
+
+// return config for tsdown
+export default defineConfig({
+  ...defaultConfig,
+  entry: [
+    './src/index.ts',
+    './src/Providers/**/*.ts',
+  ],
+});
