+++ packages/storage/src/Types/StorageTypes.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-empty-object-type */
 import type { Storage } from '../Service/Storage';
 import type { IOC } from '@vercube/di';
-import type { S3ClientConfig } from '@aws-sdk/client-s3';
 
 export namespace StorageTypes {
   export interface BaseOptions {
@@ -47,9 +46,5 @@ export namespace StorageTypes {
   }
 
   export interface Size extends BaseOptions {
-  }
-
-  export interface S3BaseOptions extends S3ClientConfig {
-    bucket: string;
   }
 }
