+++ packages/storage/src/index.ts
@@ -2,9 +2,5 @@
 export * from './Service/StorageManager';
 export * from './Service/Storage';
 
-// Strategies
-export * from './Storages/MemoryStorage';
-export * from './Storages/S3Storage';
-
 // Types
 export * from './Types/StorageTypes';
