+++ packages/storage/test/StorageManager.test.ts
@@ -1,7 +1,9 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { Container } from '@vercube/di';
 import { Logger } from '@vercube/logger';
-import { StorageManager, MemoryStorage, Storage, S3Storage } from '../src';
+import { StorageManager, Storage } from '../src';
+import { MemoryStorage } from '../src/Storages/MemoryStorage';
+import { S3Storage } from '../src/Storages/S3Storage';
 import { initializeContainer } from '@vercube/di';
 import { TestStorage, ErrorStorage } from './Utils/Mock.mock';
