+++ playground/src/Boot/Container.ts
@@ -2,7 +2,8 @@ import { Container } from '@vercube/di';
 import { ErrorHandlerProvider } from '@vercube/core';
 import { Logger } from '@vercube/logger';
 import { ConsoleProvider } from '@vercube/logger/providers/ConsoleProvider';
-import { StorageManager, MemoryStorage } from '@vercube/storage';
+import { StorageManager } from '@vercube/storage';
+import { MemoryStorage } from '@vercube/storage/storages/MemoryStorage';
 import PlaygroundController from '../Controllers/PlaygroundController';
 import { BasicAuthenticationProvider } from '../Services/BasicAuthenticationProvider';
 import { DummyAuthorizationProvider } from '../Services/DummyAuthorizationProvider';
